Harmonize iterator prototype intrinsics to follow spec
