+++ Jint/Native/Map/MapInstance.cs
@@ -85,17 +85,17 @@ namespace Jint.Native.Map
 
         internal ObjectInstance Iterator()
         {
-            return _realm.Intrinsics.Iterator.ConstructEntryIterator(this);
+            return _realm.Intrinsics.MapIteratorPrototype.ConstructEntryIterator(this);
         }
 
         internal ObjectInstance Keys()
         {
-            return _realm.Intrinsics.Iterator.ConstructKeyIterator(this);
+            return _realm.Intrinsics.MapIteratorPrototype.ConstructKeyIterator(this);
         }
 
         internal ObjectInstance Values()
         {
-            return _realm.Intrinsics.Iterator.ConstructValueIterator(this);
+            return _realm.Intrinsics.MapIteratorPrototype.ConstructValueIterator(this);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
